pyproject.toml
[tool.poetry]
name = "semaphore-client-cli"
version = "0.1.0"
description = ""
authors = ["Quien Sabe <qs5779@mail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
semaphore-client = {url = "https://privy.wtforg.net/wheels/semaphore_client-0.2.2.dev0-py3-none-any.whl"}
loguru = "^0.7.2"
pydantic-settings = "^2.3.4"
wtforglib = "^1.0.2"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
isort = "^5.13.2"
black = "^24.4.2"
wemake-python-styleguide = "^0.19.2"
mypy = "^1.10.1"
pytest-cov = "^5.0.0"
flake8 = "^7.1.0"
doc8 = "^1.1.1"
safety = "^3.2.4"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = ">=5.2,<8.0"
sphinx-autodoc-typehints = ">=1.24,<3.0"
m2r2 = "^0.3"
tomli = "^2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
